fixed data saving and added argparse
